Move exporter funcs' imports to new file
