+++ src/pages/List/ListExtras/HtmlListText.js
@@ -1,121 +1,7 @@
 import React from 'react';
 import _ from 'lodash';
 import cards from 'constants/cards';
-import ranks from 'constants/ranks';
-import legionModes from 'constants/legionModes';
-import interactions from 'constants/cardInteractions';
-import listTemplate from 'constants/listTemplate';
 
-function generateHTMLText(
-  list, showPoints = true, showCommands = false, showBattles = false
-) {
-  let header = `${list.title ? list.title : 'Untitled'}<br>`;
-  header += `${list.pointTotal}/${legionModes[list.mode].maxPoints}<br>`;
-  let units = '';
-  list.units.forEach(unit => {
-    const unitCard = cards[unit.unitId];
-    if (unit.count === 1) {
-      if (unit.unitId === 'pz') { // Kraken
-        units += `${unitCard.cardName} - Kraken (${unit.totalUnitCost})<br>`;
-      } else if (unit.unitId === 'pz') { // Kalani
-        units += `${unitCard.cardName} - Kalani (${unit.totalUnitCost})<br>`;
-      } else {
-        units += `${unitCard.cardName} (${unit.totalUnitCost})<br>`;
-      }
-      for (let j = 0; j < unit.upgradesEquipped.length; j++) {
-        if (unit.upgradesEquipped[j]) {
-          const upgradeCard = cards[unit.upgradesEquipped[j]];
-          if (unit.loadoutUpgrades && unit.loadoutUpgrades[j]) {
-            const loadoutCard = cards[unit.loadoutUpgrades[j]];
-            units += ` - ${upgradeCard.cardName} (${upgradeCard.cost})`;
-            units += `/${loadoutCard.cardName} (${loadoutCard.cost})<br>`;
-          } else {
-            units += ` - ${upgradeCard.cardName} (${upgradeCard.cost})<br>`;
-          }
-        }
-      }
-      if (unit.counterpart) {
-        const { counterpart } = unit;
-        const counterpartCard = cards[counterpart.counterpartId];
-        units += `${counterpartCard.cardName} (${unit.counterpart.totalUnitCost})\n`;
-        for (let j = 0; j < counterpart.upgradesEquipped.length; j++) {
-          if (counterpart.upgradesEquipped[j]) {
-            const upgradeCard = cards[counterpart.upgradesEquipped[j]];
-            if (counterpart.loadoutUpgrades && counterpart.loadoutUpgrades[j]) {
-              const loadoutCard = cards[counterpart.loadoutUpgrades[j]];
-              units += ` - ${upgradeCard.cardName} (${upgradeCard.cost})`;
-              units += `/${loadoutCard.cardName} (${loadoutCard.cost})<br>`;
-            } else {
-              units += ` - ${upgradeCard.cardName} (${upgradeCard.cost})<br>`;
-            }
-          }
-        }
-      }
-    } else {
-      for (let i = 0; i < unit.count; i++) {
-        units += `${unitCard.cardName} (${unit.totalUnitCost / unit.count})<br>`;
-        for (let j = 0; j < unit.upgradesEquipped.length; j++) {
-          if (unit.upgradesEquipped[j]) {
-            const upgradeCard = cards[unit.upgradesEquipped[j]];
-            units += ` - ${upgradeCard.cardName} (${upgradeCard.cost})<br>`;
-          }
-        }
-      }
-    }
-  });
-  let commands = '';
-  list.commandCards.forEach(commandId => {
-    let pips = '••••';
-    const commandCard = cards[commandId];
-    if (commandCard.cardSubtype === '1') pips = '•';
-    else if (commandCard.cardSubtype === '2') pips = '••';
-    else if (commandCard.cardSubtype === '3') pips = '•••';
-    commands += `${pips}${commandCard.cardName}<br>`;
-  });
-  if (commands !== '') {
-    commands = `<br>Commands:<br>${commands}`;
-    commands += '••••Standing Orders<br>';
-  }
-  let contingencies = '';
-  if (list.contingencies && list.contingencies.length > 0) {
-    contingencies = '<br>Contingencies:<br>';
-    list.contingencies.forEach(commandId => {
-      let pips = '••••';
-      const commandCard = cards[commandId];
-      if (commandCard.cardSubtype === '1') pips = '•';
-      else if (commandCard.cardSubtype === '2') pips = '••';
-      else if (commandCard.cardSubtype === '3') pips = '•••';
-      contingencies += `${pips}${commandCard.cardName}<br>`;
-    });
-  }
-  let primaries = '';
-  let secondaries = '';
-  let advantages = '';
-  if (list.primaryCards.length > 0) {
-    primaries += 'Primaries:<br>';
-    list.primaryCards.forEach((id, i) => {
-      const card = cards[id];
-      primaries += ` - ${card.cardName}<br>`;
-    });
-  }
-  if (list.secondaryCards.length > 0) {
-    secondaries += 'Secondaries:<br>';
-    list.secondaryCards.forEach((id, i) => {
-      const card = cards[id];
-      secondaries += ` - ${card.cardName}<br>`;
-    });
-  }
-  if (list.advantageCards.length > 0) {
-    advantages += 'Advantages:<br>';
-    list.advantageCards.forEach((id, i) => {
-      const card = cards[id];
-      advantages += ` - ${card.cardName}<br>`;
-    });
-  }
-  let battle = '';
-  if (primaries + secondaries + advantages !== '') battle = `<br>Battle Deck<br>`;
-  return '<html><p>' + header + units + commands + contingencies + battle + primaries + secondaries + advantages + '</p></html>';
-}
 
 function generateUnitComponent(unit, index) {
   let unitText = [];
